backend/requirements.txt
argcomplete==3.6.2
asgiref==3.8.1
attrs==25.4.0
certifi==2024.8.30
click==8.1.8
colorama==0.4.6
distlib==0.3.8
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.0
filelock==3.16.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==24.2
pipenv==2024.0.2
pipx==1.7.1
platformdirs==4.3.3
pluggy==1.5.0
Pygments==2.19.1
PyJWT==2.10.1
pyproject_hooks==1.2.0
pytest==8.3.4
pytz==2025.1
pywin32-ctypes==0.2.3
PyYAML==6.0.3
referencing==0.36.2
rfc3986==2.0.0
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.2.0
urllib3==2.3.0
userpath==1.9.2
virtualenv==20.26.4
zipp==3.21.0

# Additional packages for Docker deployment
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
whitenoise==6.6.0
